refactor: Move rough stuff outside framebox and independent from SILE
