+++ packages/framebox/graphics/rough.lua
@@ -1,4 +1,4 @@
-local RoughGenerator = require("packages.framebox.rough-lua.generator").RoughGenerator
+local RoughGenerator = require("rough-lua.rough.generator").RoughGenerator
 
 SU.warn("The rough.lua module is deprecated. Please use the rough-lua module instead.")
