+++ rough-lua/rough/renderer.lua
@@ -8,7 +8,7 @@
 -- Copyright (c) 2019 Preet Shihn
 --
 
-local jsshims = require("packages.framebox.rough-lua.jsshims")
+local jsshims = require("rough-lua.rough.jsshims")
 local array_concat = jsshims.array_concat
 
 local PRNG = require("packages.framebox.graphics.prng")
@@ -18,11 +18,11 @@ local prng = PRNG()
 -- I ported path-data-parser but haven't tested it for now
 -- local pathDataParser = require("packages.framebox.path-data-parser")
 -- local parsePath, normalize, absolutize = pathDataParser.parsePath, pathDataParser.normalize, pathDataParser.absolutize
-local normalize = function () SU.error("Not yet implemented") end
-local absolutize = function () SU.error("Not yet implemented") end
-local parsePath = function () SU.error("Not yet implemented") end
+local normalize = function () error("Not yet implemented") end
+local absolutize = function () error("Not yet implemented") end
+local parsePath = function () error("Not yet implemented") end
 
-local getFiller = require("packages.framebox.rough-lua.fillers.filler").getFiller
+local getFiller = require("rough-lua.rough.fillers.filler").getFiller
 
 local function cloneOptionsAlterSeed (o)
   -- PORTING NOTE:
